+++ src/Pages/Home.tsx
@@ -1,8 +1,54 @@
-import React from "react";
+import { useEffect, useState } from "react";
+import styled from "styled-components";
+import blogListDummy from "../API/DummyData/BlogList.json";
+import BlogPost from "../Components/Common/BlogPost";
 import "./Style/Home.css";
 
+//PostData Type 미리 지정
+export interface PostProps {
+  memberId: number;
+  blogId: number;
+  titleImageUrl: string;
+  blogTitle: string;
+  blogPreview: string;
+  createdAt: string;
+  writer: string;
+  commentCount: number;
+  likeCount: number;
+}
+
+export interface ArrPostProps extends Array<PostProps> {}
+
+const Sort = styled.li<{ borderBtm: boolean }>`
+  color: var(--fc-500);
+  font-weight: var(--fw-bold);
+  border-bottom: ${(props) =>
+    props.borderBtm ? "1px solid var(--fc-500)" : "none"};
+  cursor: pointer;
+`;
+
 const Home = () => {
-  return <div className="Home_Wrapper"></div>;
+  const sortArr: string[] = ["추천순", "최신순", "내 추천"];
+  const [isSortActive, setIsSortActive] = useState(0);
+
+ 
+  return (
+    <div className="Home_Wrapper">
+      <div className="Home_Container">
+        <ul className="HomeSort_Container">
+          {sortArr.map((sort, idx) => (
+            <Sort
+              key={idx}
+              borderBtm={idx === isSortActive}
+              onClick={() => setIsSortActive(idx)}
+            >
+              {sort}
+            </Sort>
+          ))}
+        </ul>
+      </div>
+    </div>
+  );
 };
 
 export default Home;
