+++ src/Pages/Style/Home.css
@@ -1,9 +1,17 @@
 .Home_Wrapper {
+  display: flex;
   width: 100%;
-  height: 100vh;
-  background-color: antiquewhite;
 }
 
 .Home_Container {
   display: flex;
+  width: 100%;
+  flex-direction: column;
+  padding: 3em 0;
+}
+
+.HomeSort_Container {
+  display: flex;
+  justify-content: space-between;
+  width: 15%;
 }
